requirements.txt
click==8.1.3
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
python-frontmatter==1.0.0
